deletes unused instantiationError and moves useInstantiationUrlReset to taskStoreContext
