+++ src/features/routing/AppRoutingContext.tsx
@@ -1,11 +1,9 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { matchPath, useLocation, useNavigate as useNativeNavigate } from 'react-router-dom';
 import type { PropsWithChildren, RefObject } from 'react';
 import type { NavigateOptions } from 'react-router-dom';
 
-import { useQueryClient } from '@tanstack/react-query';
 import { createStore } from 'zustand';
-import type { QueryClient } from '@tanstack/react-query';
 
 import { createZustandContext } from 'src/core/contexts/zustandContext';
 
@@ -78,7 +76,6 @@ export function AppRoutingProvider({ children }: PropsWithChildren) {
   // eslint-disable-next-line react-hooks/rules-of-hooks
   const location = window.inUnitTest ? useLocation() : undefined;
   const initialLocation = location ? location.pathname + location.search : undefined;
-  useInstantiationUrlReset();
 
   return (
     <Provider initialLocation={initialLocation}>
@@ -231,31 +228,4 @@ function UpdateNavigate() {
   };
 
   return null;
-}
-
-/**
- * When the URL changes (possibly because we redirected to the instance or because the user went
- * back to party selection after getting an error), we should clear the instantiation. We do this so that we're ready
- * for a possible next instantiation (when the user comes back to try instantiation again). This is needed
- * because the components that trigger instantiation might do so repeatedly - and we need to stop that.
- *
- * Some places instantiate as a direct result of a user action (clicking a button). Those will force
- * re-instantiation anyway and won't care if the previous instantiation was cleared before trying again.
- */
-export function useInstantiationUrlReset() {
-  const location = useLocation();
-  const queryClient = useQueryClient();
-
-  useEffect(() => {
-    // Reset instantiation mutations when navigating away from instance pages
-    // This prevents duplicate instantiation attempts
-    if (!location.pathname.includes('/instance/')) {
-      removeMutations(queryClient);
-    }
-  }, [location.pathname, queryClient]);
-}
-
-function removeMutations(queryClient: QueryClient) {
-  const mutations = queryClient.getMutationCache().findAll({ mutationKey: ['instantiate'] });
-  mutations.forEach((mutation) => queryClient.getMutationCache().remove(mutation));
 }
